+++ packages/adapters/trueforge/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @artemiskit/adapter-trueforge
 
+## 0.1.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @artemiskit/core@0.4.1
+
 ## 0.1.1
 
 ### Patch Changes