+++ packages/adapters/vercel-ai/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @artemiskit/adapter-vercel-ai
 
+## 0.1.14
+
+### Patch Changes
+
+- Updated dependencies
+  - @artemiskit/core@0.4.1
+
 ## 0.1.13
 
 ### Patch Changes