+++ packages/redteam/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @artemiskit/redteam
 
+## 0.3.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @artemiskit/core@0.4.1
+
 ## 0.3.1
 
 ### Patch Changes